requirements.txt
dnspython
python-dotenv
dnstwist[full]